requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
Pillow>=10.0.0
requests>=2.31.0
textblob>=0.17.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
plotly>=5.0.0